aider: Corrected HTML for download link in `templates\index.html` to use the correct SVG icon and trigger image download.
# Aider chat conversation:

ASSISTANT: To address the issue with the download button, we need to make changes to the `templates\index.html` file. We will correct the HTML for the download link for the rest of the images in the history to match the SVG used for the latest image, and ensure that the download attribute is set correctly to trigger a download rather than navigating to the image URL.

Here is the *SEARCH/REPLACE* block for the necessary changes:

templates\index.html
<source>html
<<<<<<< SEARCH
<a href="{{ image.url }}" download class="download-link">Download</a>
=======
<a href="{{ image.url }}" download class="download-link">
    <!-- Correct SVG for download icon -->
    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" class="icon-sm text-black"><path fill-rule="evenodd" clip-rule="evenodd" d="M7.70711 10.2929C7.31658 9.90237 6.68342 9.90237 6.29289 10.2929C5.90237 10.6834 5.90237 11.3166 6.29289 11.7071L11.2929 16.7071C11.6834 17.0976 12.3166 17.0976 12.7071 16.7071L17.7071 11.7071C18.0976 11.3166 18.0976 10.6834 17.7071 10.2929C17.3166 9.90237 16.6834 9.90237 16.2929 10.2929L13 13.5858L13 4C13 3.44771 12.5523 3 12 3C11.4477 3 11 3.44771 11 4L11 13.5858L7.70711 10.2929ZM5 19C4.44772 19 4 19.4477 4 20C4 20.5523 4.44772 21 5 21H19C19.5523 21 20 20.5523 20 20C20 19.4477 19.5523 19 19 19L5 19Z" fill="currentColor"></path></svg>
</a>
>>>>>>> REPLACE
</source>

This change will ensure that all download links in the image history use the correct SVG icon and that clicking the link will trigger the download of the image.